fixing erroneous colors in the logo

+++ src/detection/zpool/zpool.h
@@ -9,7 +9,7 @@ typedef struct FFZpoolResult
     uint64_t used;
     uint64_t total;
     uint64_t version;
-    uint64_t fragmentation;
+    double fragmentation;
 } FFZpoolResult;
 
 const char* ffDetectZpool(FFlist* result /* list of FFZpoolResult */);
